backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.119.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
oauthlib==3.3.1
openai==2.6.0
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
